+++ quantify_scheduler/operations/acquisition_library.py
@@ -10,7 +10,16 @@ from quantify_scheduler import Operation
 from quantify_scheduler.enums import BinMode
 
 
-class Trace(Operation):
+class AcquisitionOperation(Operation):  # pylint: disable=too-many-ancestors
+    """
+    This class is used to help differentiate an acquisition operation from the regular
+    operations. This enables us to use
+    :func:`~.visualization.pulse_diagram.plot_acquisition_operations` to highlight
+    acquisition pulses in the pulse diagrams.
+    """
+
+
+class Trace(AcquisitionOperation):  # pylint: disable=too-many-ancestors
     """The Trace acquisition protocol measures a signal s(t)."""
 
     def __init__(
@@ -94,7 +103,9 @@ class Trace(Operation):
         return self._get_signature(acq_info)
 
 
-class WeightedIntegratedComplex(Operation):
+class WeightedIntegratedComplex(
+    AcquisitionOperation
+):  # pylint: disable=too-many-ancestors
     """
     Weighted integration acquisition protocol on a
     complex signal in a custom complex window.
@@ -208,7 +219,12 @@ class WeightedIntegratedComplex(Operation):
         return self._get_signature(acq_info)
 
 
-class SSBIntegrationComplex(Operation):
+class SSBIntegrationComplex(AcquisitionOperation):  # pylint: disable=too-many-ancestors
+    """
+    This class implements a SingleSideBand Integration acquisition protocol with
+    complex results.
+    """
+
     def __init__(
         self,
         port: str,
@@ -315,7 +331,14 @@ class SSBIntegrationComplex(Operation):
         return self._get_signature(acq_info)
 
 
-class NumericalWeightedIntegrationComplex(WeightedIntegratedComplex):
+class NumericalWeightedIntegrationComplex(
+    WeightedIntegratedComplex
+):  # pylint: disable=too-many-ancestors
+    """
+    Implements a WeightedIntegratedComplex class using parameterized waveforms and
+    interpolation as the integration weights.
+    """
+
     def __init__(
         self,
         weights_a: List[complex],